Fix search validation location
